New serialization approach
Instead of inserting the serialized string into the database,
convert the 'value' in to an "integer string" using 'toString'
and insert that into the database.

'toString' basically works as follows:

data --> bytes (via serialize)
bytes --> integer
integer --> string (with ":" separator)

'toObject' works in reverse.

This results in some inflation of the data by at least 2x.